Rename a file, add deprecation warning for --load-in-4bit
